chore: clean for testnet
