+++ src/contexts/crossSubnetMessagesGraph.ts
@@ -1,0 +1,5 @@
+import { createContext } from 'react'
+
+export const CrossSubnetMessagesGraphContext = createContext<{
+  subnetsLatestBlockNumbers?: Map<string, number>
+}>({})
